+++ EssentialFeed/EssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,15 @@
+//
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoaderTask {
+	func cancel()
+}
+
+public protocol FeedImageDataLoader {
+	typealias Result = Swift.Result<Data, Error>
+	
+	func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
